common.hmmer, common.pfamdb: fix cyclic import
