+++ antismash/common/hmmer.py
@@ -9,11 +9,11 @@ import logging
 import os
 from typing import Any, Dict, Iterable, List, Optional, Set, Tuple
 
-from antismash.common import fasta, module_results, path, pfamdb, subprocessing
+from antismash.common import fasta, module_results, pfamdb, subprocessing
+from antismash.common.pfamdb import ensure_database_pressed  # used by others # pylint: disable=unused-import
 from antismash.common.secmet import Record, CDSFeature
 from antismash.common.secmet.features import FeatureLocation, PFAMDomain
 from antismash.common.secmet.locations import location_from_string
-from antismash.config import get_config
 
 
 @dataclass(frozen=True)
@@ -277,39 +277,3 @@ def run_hmmer(record: Record, features: Iterable[CDSFeature], max_evalue: float,
         for locus_hits in results_by_cds.values():
             hits.extend(remove_overlapping(locus_hits, cutoffs))
     return HmmerResults(record.id, max_evalue, min_score, database, tool, hits)
-
-
-def ensure_database_pressed(filepath: str, return_not_raise: bool = False) -> List[str]:
-    """ Ensures that the given HMMer database exists and that the hmmpress
-        generated files aren't out of date.
-
-        Arguments:
-            filepath: the path to the HMMer database
-            return_not_raise: whether to catch errors and return their messages as strings
-
-        Returns:
-            any encountered error messages, will never be populated without return_not_raise == True
-    """
-    components = [f"{filepath}.{ext}" for ext in ["h3f", "h3i", "h3m", "h3p"]]
-
-    if "mounted_at_runtime" in filepath:
-        msg = f"Cannot ensure database pressed when set to mount at runtime: {filepath}"
-        if return_not_raise:
-            return [msg]
-        raise ValueError(msg)
-
-    if path.is_outdated(components, filepath):
-        logging.info("%s components missing or obsolete, re-pressing database", filepath)
-        if "hmmpress" not in get_config().executables:
-            msg = f"Failed to hmmpress {filepath!r}: cannot find executable for hmmpress"
-            if not return_not_raise:
-                raise RuntimeError(msg)
-            return [msg]
-
-        result = subprocessing.run_hmmpress(filepath)
-        if not result.successful():
-            msg = f"Failed to hmmpress {filepath!r}: {result.stderr}"
-            if not return_not_raise:
-                raise RuntimeError(msg)
-            return [msg]
-    return []
